requirements.txt
# Core Dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Computer Vision & Deep Learning
torch>=1.9.0
torchvision>=0.10.0
open-clip-torch>=2.0.0
Pillow>=8.0.0

# API Integration
openai>=1.0.0

# Statistical Analysis
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0

# Optional: Development
# jupyter>=1.0.0
# notebook>=6.4.0
# pytest>=6.2.0
